Remove direct dependency on aes crate
In practice everyone will use FF1 with AES, as that is the only block
cipher that fits the necessary NIST requirements. But this crate does
not need to depend specifically on the aes crate for anything other
than tests. This commit enables downstream users to use a different
AES implementation if desired (e.g. in an embedded environment with
an existing AES implementation).